elt/requirements.txt
# Database connectors
pymysql==1.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Data processing
pandas==2.2.2
numpy==1.26.4

# Environment
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Scheduling (optional)
schedule==1.2.1